fix: making example and docs less confusing (#5281)
