+++ examples/custom_search_method/asha_search_method/local_search_runner/run_experiment.py
@@ -20,12 +20,39 @@ import sys
 sys.path.append(".")
 
 import logging
+import determined as det
 from asha import ASHASearchMethod
-from utils import sample_params
+from typing import Dict
+import random
 from pathlib import Path
 from determined import searcher
 
+
+############################################################################
+# User-defined function that generates a combination of hyperparameters for each trial.
+# The hyperparameters are passed to a trial in the Create operation.
+# In this example, the model (defined in experiment_files/model_def.py) is expecting
+# the following hyperparameters:
+#   -> global_batch_size,
+#   -> n_filters1
+#   -> n_filters2,
+#   -> learning_rate,
+#   -> dropout,
+#   -> dropout2.
+def sample_params() -> Dict[str, object]:
+    hparams = {
+        "global_batch_size": 64,
+        "n_filters1": random.randint(8, 64),
+        "n_filters2": random.randint(8, 72),
+        "learning_rate": 10 ** random.uniform(-4.0, 0.0),
+        "dropout1": random.uniform(0.2, 0.8),
+        "dropout2": random.uniform(0.2, 0.8),
+    }
+    return hparams
+
+
 if __name__ == "__main__":
+    logging.basicConfig(level=logging.INFO, format=det.LOG_FORMAT)
 
     ########################################################################
     # Multi-trial experiment
