+++ examples/custom_search_method/asha_search_method/remote_search_runner/searcher.yaml
@@ -1,14 +1,9 @@
-name: core-search-runner
+name: remote-search-runner
 entrypoint: python3 remote_search_runner/run_experiment.py
-hyperparameters:
-  max_length: 1000
-  max_trials: 16
-  num_rungs: 3
-  divisor: 4
 searcher:
   metric: validation_error
   smaller_is_better: true
   name: single
   max_length:
-    batches: 100
+    batches: 1000
 max_restarts: 0
